pyproject.toml
[project]
name = "flow-matching"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "aim>=3.29.1",
    "datasets>=4.3.0",
    "diffusers>=0.35.2",
    "hydra-core>=1.3.2",
    "isort>=7.0.0",
    "numpy>=2.3.4",
    "nvitop>=1.5.3",
    "protobuf<4.0.0",
    "py-spy>=0.4.1",
    "ruff>=0.14.3",
    "sentencepiece>=0.2.1",
    "torch>=2.9.0",
    "torchinfo>=1.8.0",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "wandb>=0.22.3",
    "webdataset>=1.0.2",
]

[tool.pyright]
reportMissingTypeStubs = false

[tool.ruff]
select = ["I"] # Enable isort rules

[tool.ruff.isort]
# Common isort-style configurations
known-first-party = ["flow_matching"]
known-third-party = ["numpy", "pandas"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]
lines-after-imports = 2
force-single-line = false
force-sort-within-sections = false

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }
